requirements.txt
aiogram~=3.19.0
loguru
python-dotenv
aiohttp
aiosqlite
dataclasses-json
rich
apscheduler